+++ spec/rubocop/cop/style/next_spec.rb
@@ -28,7 +28,7 @@ describe RuboCop::Cop::Style::Next, :config do
                                             'end'])
       expect(new_source).to eq(['3.downto(1) do',
                                 "  next #{opposite} o == 1",
-                                '    puts o', # another cop can fix indentation
+                                '  puts o',
                                 'end'].join("\n"))
     end
 
@@ -51,7 +51,7 @@ describe RuboCop::Cop::Style::Next, :config do
                                             'end'])
       expect(new_source).to eq(['[].each do |o|',
                                 "  next #{opposite} o == 1",
-                                '    puts o', # another cop can fix indentation
+                                '  puts o',
                                 'end'].join("\n"))
     end
 
@@ -85,7 +85,7 @@ describe RuboCop::Cop::Style::Next, :config do
                                             'end'])
       expect(new_source).to eq(['for o in 1..3 do',
                                 "  next #{opposite} o == 1",
-                                '    puts o', # another cop can fix indentation
+                                '  puts o',
                                 'end'].join("\n"))
     end
 
@@ -317,11 +317,31 @@ describe RuboCop::Cop::Style::Next, :config do
                                           'end'])
     expect(new_source).to eq(['loop do',
                               '  next unless test # keep me',
-                              '    # keep me',
-                              '    something # keep me',
-                              '    # keep me',
+                              '  # keep me',
+                              '  something # keep me',
+                              '  # keep me',
                               '    ',
                               ' # keep me',
+                              'end'].join("\n"))
+  end
+
+  it "doesn't reindent heredoc bodies when autocorrecting" do
+    new_source = autocorrect_source(cop, ['loop do',
+                                          '  if test',
+                                          '    str = <<-BLAH',
+                                          '  this is a heredoc',
+                                          '   nice eh?',
+                                          '    BLAH',
+                                          '    something',
+                                          '  end',
+                                          'end'])
+    expect(new_source).to eq(['loop do',
+                              '  next unless test',
+                              '  str = <<-BLAH',
+                              '  this is a heredoc',
+                              '   nice eh?',
+                              '  BLAH',
+                              '  something',
                               'end'].join("\n"))
   end
 
@@ -337,11 +357,10 @@ describe RuboCop::Cop::Style::Next, :config do
                                           'end'])
     expect(new_source).to eq(['loop do',
                               '  next unless test',
-                              '    loop do', # another cop can fix indentation
-                              '      next unless test',
-                              # another cop can fix indentation
-                              '        something',
-                              '    end',
+                              '  loop do',
+                              '    next unless test',
+                              '    something',
+                              '  end',
                               'end'].join("\n"))
   end